+++ tests/core/test_thunks.py
@@ -2,11 +2,12 @@
 
 import pytest
 
-from aioway._core import TensorInput, Thunk, TorchThunk
+from aioway._core import TensorInput, Thunk
 from aioway.modes import (
     Aten,
     AtenThunk,
     HistTensorGraph,
+    ModeThunk,
     NnFwdThunk,
     NnInitThunk,
     TorchDispThunk,
@@ -26,7 +27,7 @@ def _fn_cls():
 
 def _input_cls():
     yield HistTensorGraph
-    yield TorchThunk
+    yield ModeThunk
     yield AtenThunk
     yield NnFwdThunk
     yield AtenThunk
